requirements.txt
discord.py-self
requests
beautifulsoup4
python-dotenv
aiohttp
pytz
yfinance
aiosqlite
